producer/requirements.txt
websocket-client
pymongo
kafka-python
